Clean up the storage
