+++ src/containers/storage.tsx
@@ -11,30 +11,18 @@ import Column from '../components/layout/column';
 import StorageTypeViewer from "../components/viewer/StorageTypeViewer"
 import { StorageState, StorageRecord } from '../store/reducers/storageReducer';
 import StorageRecordViewer from '../components/viewer/StorageRecordViewer';
+import { sortByDate } from '../sort/sort';
+
 
 const Storage = () => {
   const dispatch = useDispatch();
   const crops = useSelector<RootState, CropState["crops"]>((state) => state.crops.crops);
   const records = useSelector<RootState, StorageState["records"]>((state) => state.storage.records);
+  let sortedRecords = sortByDate(records);
 
-  const sortByDate = (records: { date: string }[]) => {
-    const sorted = records.filter(() => true).sort((recordA, recordB) => {
-      const dateA = dateToYear(recordA.date);
-      const dateB = dateToYear(recordB.date);
-
-      return dateA - dateB;
-    })
-
-    return sorted;
-  }
-
-  const dateToYear = (date: string) => {
-    const [days, months, years] = date.split("/");
-    const dayToYear = 1 / 360;
-    const monthToYear = 1 / 12;
-
-    return Number(years) + Number(months) * monthToYear + Number(days) * dayToYear;
-  }
+  useEffect(() => {
+    sortedRecords = sortByDate(records);
+  })
 
   const onRemoveClick = (id: number) => {
     dispatch(removeStorageRecord(id));
@@ -61,11 +49,19 @@ const Storage = () => {
   }
 
   const calcProductAmount = (product: string) => {
-    let sorted = records.filter(record => record.product === product);
-    sorted = sortByDate(sorted) as StorageRecord[];
+    let records = leaveOnlyRecordsBelongTo(product);
+
+    return totalRecordsUp(records);
+  }
+
+  const leaveOnlyRecordsBelongTo = (product: string) => {
+    return sortedRecords.filter(record => record.product === product);
+  }
+
+  const totalRecordsUp = (records: StorageRecord[]) => {
     let result = 0;
 
-    for (const { amount, type } of sorted) {
+    for (const { amount, type } of records) {
       if (type === "Przychud")
         result += amount;
       else
@@ -76,35 +72,24 @@ const Storage = () => {
   }
 
   const calcProductAmountForRecord = (product: string, record: StorageRecord) => {
-    let result = 0;
-    let sorted = records.filter(record => record.product === product);
+    let records = leaveOnlyRecordsBelongTo(product);
+    records = removeRecordAfter(records, record);
 
-    sorted = sortByDate(sorted) as StorageRecord[];
-    console.log(sorted);
-
-    const lastId = sorted.indexOf(record);
-    console.log(lastId);
-
-    const finalRecords = sorted.splice(0, lastId + 1);
-    console.log(finalRecords);
-
-    for (const { amount, type } of finalRecords) {
-      if (type === "Przychud")
-        result += amount;
-      else
-        result -= amount;
-    }
-
-    return result
+    return totalRecordsUp(records);
   }
 
-  const products = (crops: Crop[]) => {
+  const removeRecordAfter = (records:StorageRecord[], record: StorageRecord) => {
+    const limit = records.indexOf(record);
+    
+    return records.splice(0, limit + 1);
+  }
+
+  const getProducts = (crops: Crop[]) => {
     const products = crops.map(crop => crop.plant + " " + crop.plantVariant);
     const uniqueProducts = Array.from(new Set(products));
 
     return uniqueProducts;
   }
-
 
   return (
     <>
@@ -113,7 +98,7 @@ const Storage = () => {
       </Flex>
 
       <Flex width="100%">
-        {products(crops).map(product =>
+        {getProducts(crops).map(product =>
           <React.Fragment key={product}>
             <Column width="297px">
               <StorageTypeViewer amount={calcProductAmount(product)}>
